Rename `bitstream_status_list` into `bitstring_status_list`.
