+++ crates/status/src/impl/mod.rs
@@ -1,5 +1,5 @@
 pub mod any;
-pub mod bitstream_status_list;
+pub mod bitstring_status_list;
 pub mod token_status_list;
 
 pub use flate2::Compression;
